Move some config function to a helper file for better visibility
